Split itertools.hpp into multiple header files
